Move JSONDict
