+++ datadog_checks_dev/setup.py
@@ -25,7 +25,6 @@ REQUIRES = [
     "contextlib2; python_version < '3.0'",
     'coverage>=5.0.3',
     'flaky',
-    'jsonpointer',
     'mock',
     'psutil',
     'PyYAML==5.4.1',
@@ -81,6 +80,7 @@ setup(
             'datamodel-code-generator~=0.11.4',
             'docker-compose>=1.25',
             'in-toto>=0.4.2',
+            'jsonpointer',
             'jsonschema',
             'markdown',
             'orjson',
